server now also successfully runs tests
